-rework on CallBack code for the notification of occured events -introducing namespace in order to structure the middleware accordingly  and to allow better overview and forbid name clashes -adapting file properties
git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@471 235a714a-1e17-0410-b3bb-947d69a96898

+++ include/mw/common/UID.h
@@ -5,11 +5,12 @@
 
 struct UID {
     union {
-	uint64_t value;
+	unsigned long long value;
+//	uint64_t value;
 	uint8_t tab[8];
     };
-    
-    UID() {} 
+
+    UID() {}
     UID(uint8_t uid)  : value(uid) {}
     UID(uint16_t uid) : value(uid) {}
     UID(uint32_t uid) : value(uid) {}
@@ -19,7 +20,7 @@ struct UID {
 	return value==a.value;
     }
 
-//     const UID& 
+//     const UID&
 };
 
 #endif /* __UID_h__ */
